Added Caching to Access Groups

+++ app/Http/Controllers/User/AccessGroupController.php
@@ -38,8 +38,14 @@ class AccessGroupController extends APIController
 	public function index()
 	{
 		if (!$this->api) return view('access.groups.index');
-		$access_groups = AccessGroup::select(['id','name'])->get();
-		return $this->reply($access_groups->pluck('name','id'));
+
+		if(env('APP_DEBUG') == "true") \Cache::forget('access_groups');
+		$access_groups = \Cache::remember('access_groups', 1800,  function () {
+			$access_groups = AccessGroup::select(['id','name'])->get();
+			return $access_groups->pluck('name','id');
+		});
+
+		return $this->reply($access_groups);
 	}
 
 	/**
@@ -75,8 +81,17 @@ class AccessGroupController extends APIController
 		$invalid = $this->validateAccessGroup($request);
 		if($invalid) return $this->setStatusCode(400)->reply($invalid);
 
-		$access_group = AccessGroup::create($request->all());
-		if (!$this->api) return redirect()->route('access.groups.show', ['group_id' => $access_group->id]);
+
+		\DB::transaction(function () use($request) {
+			$access_group = AccessGroup::create($request->only(['name','description']));
+			if($request->filesets) {
+				foreach($request->filesets as $fileset) $access_group->filesets()->create(['hash_id' => $fileset]);
+				foreach($request->users as $user) $access_group->users()->create(['user_id' => $user]);
+				foreach($request->types as $type) $access_group->users()->create(['user_id' => $user]);
+			}
+		});
+
+		//if (!$this->api) return redirect()->route('access.groups.show', ['group_id' => $access_group->id]);
 		return $this->reply(["message" => "Access Group Successfully Created"]);
 	}
 
@@ -110,8 +125,12 @@ class AccessGroupController extends APIController
 	 */
 	public function show($id)
 	{
-		$access_group = AccessGroup::with('filesets','types','keys')->find($id);
-		return $this->reply(fractal($access_group, new AccessGroupTransformer()));
+		if(env('APP_DEBUG') == "true") \Cache::forget('access_group_'.$id);
+		$access_group = \Cache::remember('access_group_'.$id, 1800,  function () use($id) {
+			$access_group = AccessGroup::with('filesets','types','users')->find($id);
+			return fractal($access_group, new AccessGroupTransformer());
+		});
+		return $this->reply($access_group);
 	}
 
 	/**
@@ -187,10 +206,11 @@ class AccessGroupController extends APIController
 	 */
 	public function destroy($id)
 	{
+		if(!$this->validateUser()) return $this->replyWithError("not authorized");
 		$access_group = AccessGroup::find($id);
 		$access_group->delete();
 
-		return $this->reply("successfully delete");
+		return $this->reply("successfully deleted");
 	}
 
 	/**
@@ -203,7 +223,7 @@ class AccessGroupController extends APIController
 	private function validateAccessGroup(Request $request)
 	{
 		$validator = \Validator::make($request->all(), [
-			'name'               => 'required|max:64|alpha_dash',
+			'name'               => ($request->method() == 'POST') ? 'required|max:64|alpha_dash|unique:dbp.access_groups,name' : 'required|max:64|alpha_dash|exists:dbp.access_groups,name',
 			'description'        => 'string',
 			'filesets.*'         => 'exists:dbp.bible_filesets,hash_id',
 			'keys.*'             => 'exists:dbp_users.user_keys,key',
